requirements.txt
# Core RAG
langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2

# PDF Processing
pypdf==4.0.0
python-docx==1.1.0
unstructured==0.12.0
tabula-py==2.7.0
camelot-py==0.11.5

# OCR (optional, for images)
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# UI
gradio==4.16.0

# Utilities
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
